+++ packages/cli/src/loggers/plainLogger.ts
@@ -2,8 +2,8 @@ import { relative } from 'node:path'
 import { formatMs, toCause } from '@internals/utils'
 import { defineLogger, logLevel as logLevelMap } from '@kubb/core'
 import { SUMMARY_SEPARATOR } from '../constants.ts'
-import { getSummary } from '../utils/getSummary.ts'
-import { runHook } from '../utils/runHook.ts'
+import { getSummary } from './utils.ts'
+import { runHook } from '../utils.ts'
 import { formatCommandWithArgs, formatMessage } from './utils.ts'
 
 /**
