+++ packages/cli/src/runners/mcp.ts
@@ -2,14 +2,29 @@ import process from 'node:process'
 import { styleText } from 'node:util'
 import { getErrorMessage } from '@internals/utils'
 import type * as McpModule from '@kubb/mcp'
-import { buildTelemetryEvent, sendTelemetry } from '../utils/telemetry.ts'
+import { buildTelemetryEvent, sendTelemetry } from '../telemetry.ts'
 
 type McpOptions = {
+  /**
+   * Current `@kubb/cli` version string, used for the telemetry payload.
+   */
   version: string
+  /**
+   * TCP port for the HTTP MCP server. When `undefined`, the server uses stdio transport.
+   */
   port?: string
+  /**
+   * Hostname to bind to when running in HTTP mode.
+   *
+   * @default 'localhost'
+   */
   host?: string
 }
 
+/**
+ * Starts the Kubb MCP server using `@kubb/mcp`.
+ * Exits the process with code 1 if `@kubb/mcp` is not installed.
+ */
 export async function runMcp({ version, port, host }: McpOptions): Promise<void> {
   let mod: typeof McpModule
   try {
@@ -29,7 +44,7 @@ export async function runMcp({ version, port, host }: McpOptions): Promise<void>
   const hrStart = process.hrtime()
   try {
     console.log('⏳ Starting MCP server...')
-    console.warn(styleText('yellow', 'This feature is still under development — use with caution'))
+    console.warn(styleText('yellow', 'This feature is still under development, use with caution'))
     run(undefined, {
       port: port !== undefined ? Number(port) : undefined,
       host,
