+++ packages/cli/src/runners/validate.ts
@@ -1,22 +1,38 @@
 import process from 'node:process'
 import { styleText } from 'node:util'
 import { getErrorMessage } from '@internals/utils'
-import { buildTelemetryEvent, sendTelemetry } from '../utils/telemetry.ts'
+import { buildTelemetryEvent, sendTelemetry } from '../telemetry.ts'
 
 type ValidateOptions = {
+  /**
+   * Path or URL to the OpenAPI/Swagger file to validate.
+   */
   input: string
+  /**
+   * Current `@kubb/cli` version string, used for the telemetry payload.
+   */
   version: string
 }
 
 type ValidateModule = typeof import('@kubb/adapter-oas')
 type ValidateDependencies = {
+  /**
+   * Loads `@kubb/adapter-oas`. Injected so tests can substitute a mock.
+   */
   loadValidateModule: () => Promise<ValidateModule>
 }
 
+/**
+ * Dynamically loads `@kubb/adapter-oas` for OpenAPI validation.
+ */
 export function loadValidateModule(): Promise<ValidateModule> {
   return import('@kubb/adapter-oas') as Promise<ValidateModule>
 }
 
+/**
+ * Validates an OpenAPI/Swagger file at `input` using `@kubb/adapter-oas`.
+ * Exits the process with code 1 on validation failure or missing dependency.
+ */
 export async function runValidate({ input, version }: ValidateOptions, dependencies: ValidateDependencies = { loadValidateModule }): Promise<void> {
   const hrStart = process.hrtime()
   try {
