+++ packages/cli/src/telemetry.test.ts
