--- packages/cli/src/utils/getCosmiConfig.ts
@@ -1,75 +1,0 @@
-import type { Config } from '@kubb/core'
-import { cosmiconfig } from 'cosmiconfig'
-import { createJiti } from 'jiti'
-
-type CosmiconfigResult = {
-  filepath: string
-  isEmpty?: boolean
-  config: Config
-}
-
-const jiti = createJiti(import.meta.url, {
-  jsx: {
-    runtime: 'automatic',
-    importSource: '@kubb/renderer-jsx',
-  },
-  moduleCache: false,
-})
-
-const tsLoader = async (configFile: string) => {
-  return jiti.import(configFile, { default: true })
-}
-
-export async function getCosmiConfig(moduleName: string, config?: string): Promise<CosmiconfigResult> {
-  let result: CosmiconfigResult
-  const searchPlaces = [
-    'package.json',
-    `.${moduleName}rc`,
-    `.${moduleName}rc.json`,
-    `.${moduleName}rc.yaml`,
-    `.${moduleName}rc.yml`,
-
-    `.${moduleName}rc.ts`,
-    `.${moduleName}rc.mts`,
-    `.${moduleName}rc.cts`,
-    `.${moduleName}rc.js`,
-    `.${moduleName}rc.mjs`,
-    `.${moduleName}rc.cjs`,
-
-    `${moduleName}.config.ts`,
-    `${moduleName}.config.mts`,
-    `${moduleName}.config.cts`,
-    `${moduleName}.config.js`,
-    `${moduleName}.config.mjs`,
-    `${moduleName}.config.cjs`,
-  ]
-  const explorer = cosmiconfig(moduleName, {
-    cache: false,
-    searchPlaces: [
-      ...searchPlaces.map((searchPlace) => {
-        return `.config/${searchPlace}`
-      }),
-      ...searchPlaces.map((searchPlace) => {
-        return `configs/${searchPlace}`
-      }),
-      ...searchPlaces,
-    ],
-    loaders: {
-      '.ts': tsLoader,
-      '.mts': tsLoader,
-      '.cts': tsLoader,
-    },
-  })
-
-  try {
-    result = config ? ((await explorer.load(config)) as CosmiconfigResult) : ((await explorer.search()) as CosmiconfigResult)
-  } catch (error) {
-    throw new Error('Config failed loading', { cause: error })
-  }
-
-  if (result?.isEmpty || !result || !result.config) {
-    throw new Error('Config not defined, create a kubb.config.js or pass through your config with the option --config')
-  }
-
-  return result as CosmiconfigResult
-}
